+++ src/java/com/squidhq/plugin/APISingleton.java
@@ -1,13 +1,11 @@
 package com.squidhq.plugin;
 
-import com.squidhq.reef.Reef;
-
 public class APISingleton {
 
     public static API _api;
 
     /**
-     * @deprecated Use {@link Reef#getAPI()} instead.
+     * @deprecated {@see <a href="https://github.com/masterdoctor/SquidHQ-Reef/wiki/Developers">https://github.com/masterdoctor/SquidHQ-Reef/wiki/Developers</a>}
      */
     public static API getAPI() {
         return _api;
