+++ src/java/com/squidhq/reef/api/placeholders/util/StockPlaceholders.java
@@ -1,4 +1,4 @@
-package com.squidhq.reef.api.placeholders;
+package com.squidhq.reef.api.placeholders.util;
 
 import org.bukkit.ChatColor;
 import org.bukkit.entity.Player;
